Feature: Dockerfile added
A short dockerfile was added in order to use the application inside of
docker containers. All the files will be copied into the image. At this
point the image is runnable but the application is instable. This is the
reason why testing is hard currently.
A database connection is not implemented yet which will result in more errors.